power: qpnp-smbcharger: Hide charge counter on selected targets
Some devices have a bug where querying the charge counter during deep
sleep takes an excessively long amount of time. BatteryService looks
up this prop during deep sleep, resulting in a very slow wakeup on these
devices.

land has this bug while santoni does not. Add a filter to hide the
charge counter on affected targets, so BatteryService does not attempt
to read the prop. It is unknown as to whether this is a firmware or a
hardware issue.

Change-Id: I099788f1db675b64575ce122b59a12035b49e529